feat: save visited pages to database
