+++ src/App.jsx
@@ -5,23 +5,19 @@ import Dashboard from './components/Dashboard/Dashboard';
 import { Routes, Route } from 'react-router-dom';
 import Settings from './components/Settings/Settings';
 
-import AuthContextProvider from './contexts/AuthContextProvider';
-
-import { mouseData } from './utils/mouseData';
+import { MouseData } from './utils/mouseData';
 
 const App = () => {
-	mouseData();
+	MouseData();
 
 	return (
-		<AuthContextProvider>
-			<div className="grid h-full grid-rows-[max-content,1fr]">
-				<Toolbar />
-				<Routes>
-					<Route path="/" element={<Dashboard />} />
-					<Route path="settings/*" element={<Settings />}></Route>
-				</Routes>
-			</div>
-		</AuthContextProvider>
+		<div className="grid h-full grid-rows-[max-content,1fr]">
+			<Toolbar />
+			<Routes>
+				<Route path="/" element={<Dashboard />} />
+				<Route path="settings/*" element={<Settings />}></Route>
+			</Routes>
+		</div>
 	);
 };
